+++ packages/cli/index.js
@@ -3,7 +3,7 @@ require('colors');
 const chalk = require('chalk');
 const path = require('path');
 const program = require('commander');
-const generateCompilation = require('./lib/compile');
+const generateCompilation = require('./lifecycles/compile');
 const runProdBuild = require('./tasks/build');
 const runDevServer = require('./tasks/develop');
 const scriptPkg = require(path.join(__dirname, '../..', 'package.json'));
