+++ packages/cli/lib/browser.js
