+++ packages/cli/lib/renderer.js
