+++ packages/cli/lifecycles/config.js
