+++ packages/cli/lifecycles/scaffold.js
