+++ packages/cli/plugins/meta.js
